refactor: moved utilities for text formatting to text-format-util.ts
