--- src/utils/getTrimmedText.ts
@@ -1,3 +1,0 @@
-const getTrimmedText = (node: Node): string | undefined => node.textContent?.trim();
-
-export default getTrimmedText;
