--- src/utils/toAsciiString.ts
@@ -1,6 +1,0 @@
-const fullWidthAscii = /[Ａ-Ｚａ-ｚ０-９]/g;
-
-const toAsciiString = (str: string) =>
-  str.replace(fullWidthAscii, (s) => String.fromCharCode(s.charCodeAt(0) - 0xfee0));
-
-export default toAsciiString;
